+++ src/main/kotlin/blackjack/domain/card/Card.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.card
 
 data class Card(val number: CardNumber, val shape: CardShape) {
