+++ src/main/kotlin/blackjack/domain/participant/Players.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.participant
 
 class Players(private val players: Set<Player>) {
     init {
