+++ src/main/kotlin/blackjack/domain/state/Betting.kt
@@ -1,8 +1,8 @@
 package blackjack.domain.state
 
-import blackjack.domain.Card
-import blackjack.domain.Hand
 import blackjack.domain.Money
+import blackjack.domain.card.Card
+import blackjack.domain.card.Hand
 
 class Betting(override val hand: Hand) : Started(hand) {
     override val bettingMoney: Money? = null
