+++ src/main/kotlin/blackjack/domain/state/Bust.kt
@@ -1,7 +1,7 @@
 package blackjack.domain.state
 
-import blackjack.domain.Hand
 import blackjack.domain.Money
+import blackjack.domain.card.Hand
 
 class Bust(override val hand: Hand, override val bettingMoney: Money?) : Finished(hand, bettingMoney) {
     override val earningRate: Double = 1.0
