+++ src/main/kotlin/blackjack/domain/state/Finished.kt
@@ -1,8 +1,8 @@
 package blackjack.domain.state
 
-import blackjack.domain.Card
-import blackjack.domain.Hand
 import blackjack.domain.Money
+import blackjack.domain.card.Card
+import blackjack.domain.card.Hand
 
 abstract class Finished(override val hand: Hand, override val bettingMoney: Money?) : State {
