+++ src/main/kotlin/blackjack/domain/state/Hit.kt
@@ -1,8 +1,8 @@
 package blackjack.domain.state
 
-import blackjack.domain.Card
-import blackjack.domain.Hand
 import blackjack.domain.Money
+import blackjack.domain.card.Card
+import blackjack.domain.card.Hand
 
 class Hit(override val hand: Hand, override val bettingMoney: Money?) : Running(hand, bettingMoney) {
     override fun draw(card: Card): State {
