+++ src/main/kotlin/blackjack/domain/state/State.kt
@@ -1,8 +1,8 @@
 package blackjack.domain.state
 
-import blackjack.domain.Card
-import blackjack.domain.Hand
 import blackjack.domain.Money
+import blackjack.domain.card.Card
+import blackjack.domain.card.Hand
 
 interface State {
     val hand: Hand
