+++ src/test/kotlin/blackjack/domain/card/HandTest.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.card
 
 import org.assertj.core.api.Assertions.*
 import org.junit.jupiter.api.Test
