+++ src/test/kotlin/blackjack/domain/participant/DealerTest.kt
@@ -1,5 +1,8 @@
-package blackjack.domain
+package blackjack.domain.participant
 
+import blackjack.domain.card.Card
+import blackjack.domain.card.CardNumber
+import blackjack.domain.card.CardShape
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
