+++ src/test/kotlin/blackjack/domain/participant/ParticipantNameTest.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.participant
 
 import org.assertj.core.api.Assertions.assertThatIllegalArgumentException
 import org.junit.jupiter.api.Test
